requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
dj-database-url==0.5.0
Django==3.1.5
gunicorn==20.1.0
h11==0.13.0
idna==3.3
MouseInfo==0.1.3
outcome==1.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
PySocks==1.7.1
pytweening==1.0.4
pytz==2021.3
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
whitenoise==6.0.0
wsproto==1.1.0
